requirements.txt
altair
bs4
numpy
pandas
pydeck
streamlit
dnspython
python-whois
requests